+++ bin/setup_commands.dart
@@ -1,8 +1,8 @@
-// ignore_for_file: avoid_print
+// ignore_for_file: avoid_print, constant_identifier_names
 
-import 'dart:convert';
 import 'dart:io';
 
+import 'package:stack_trace/stack_trace.dart';
 import 'package:archive/archive_io.dart';
 import 'package:args/command_runner.dart';
 import 'package:yaml/yaml.dart';
@@ -19,92 +19,70 @@ abstract class BaseSetupCommand extends Command {
 
   String get arch {
     final arch_ = argResults?["arch"] as String;
-    return arch_;
+    return ARCH_MAP[os]?[arch_] ?? arch_;
   }
+
+  String get pkgRoot => Frame.caller(1).uri.resolve("..").toFilePath();
 
   bool get force => argResults?.flag("force") ?? false;
   String get os => name;
 
-  Future<void> downloadAndExtract() async {
-    // Detect dependencies
-    // Assumed package root
-    final root = Directory.current.uri;
-    print('Building with assumed project root in: ${root.toFilePath()}');
+  Future<void> downloadAndExtract(String downArch, [String? extractPath]) async {
+    final opencvRoot = pkgRoot;
+    print(asInfo('Using package:$setupPkgName from $opencvRoot'));
 
-    // Assumed package_config.json
-    final packageConfigFile = File.fromUri(
-      root.resolve('.dart_tool/package_config.json'),
-    );
-    dynamic packageConfig;
-    try {
-      packageConfig = json.decode(await packageConfigFile.readAsString());
-    } on FileSystemException {
-      print('Missing .dart_tool/package_config.json');
-      print('Run `flutter pub get` first.');
-      exit(1);
-    } on FormatException {
-      print('Invalid .dart_tool/package_config.json');
-      print('Run `flutter pub get` first.');
-      exit(1);
-    }
-
-    // Determine the source path of package:webcrypto in the PUB_CACHE
-    final pkg = (packageConfig['packages'] ?? []).firstWhere(
-      (e) => e['name'] == setupPkgName,
-      orElse: () => null,
-    );
-    if (pkg == null) {
-      print('dependency on package:$setupPkgName is required');
-      exit(1);
-    }
-    final opencvRoot = packageConfigFile.uri.resolve(pkg['rootUri'] ?? '');
-    print('Using package:$setupPkgName from ${opencvRoot.toFilePath()}');
-
-    final doc = loadYaml(File("${opencvRoot.toFilePath()}/pubspec.yaml").readAsStringSync());
+    final doc = loadYaml(File(p.join(opencvRoot, "pubspec.yaml")).readAsStringSync());
     final version = doc["binary_version"] as String;
-    final libTarName = "libopencv_dart-$os-$arch.tar.gz";
+    final libTarName = "libopencv_dart-$os-$downArch.tar.gz";
 
-    String extractPath = "";
-    switch (os) {
-      case OS.windows:
-        extractPath = p.join(opencvRoot.toFilePath(), "windows");
-        break;
-      case OS.linux:
-        extractPath = p.join(opencvRoot.toFilePath(), "linux");
-        break;
-      case OS.android:
-        extractPath = p.join(opencvRoot.toFilePath(), "android", "src", "main", "jniLibs", arch);
-      case OS.macos:
-        extractPath = p.join(opencvRoot.toFilePath(), "macos");
-      case OS.ios:
-        extractPath = p.join(opencvRoot.toFilePath(), "ios");
-      case OS.fuchsia:
-      default:
-        throw UnsupportedError("Platform $os not supported");
+    if (extractPath == null) {
+      switch (os) {
+        case OS.windows:
+          extractPath = p.join(opencvRoot, "windows");
+          break;
+        case OS.linux:
+          extractPath = p.join(opencvRoot, "linux");
+          break;
+        case OS.android:
+          extractPath = p.join(opencvRoot, "android", "src", "main", "jniLibs", downArch);
+        case OS.macos:
+          extractPath = p.join(opencvRoot, "macos");
+        case OS.ios:
+          extractPath = p.join(opencvRoot, "ios");
+        case OS.fuchsia:
+        default:
+          throw UnsupportedError(asError("Platform $os not supported"));
+      }
     }
+
     if (!Directory(extractPath).existsSync()) {
       Directory(extractPath).createSync(recursive: true);
     }
 
-    final cacheTarPath = p.join(opencvRoot.toFilePath(), ".dart_tool", ".cache", libTarName);
+    final cacheTarPath = p.join(opencvRoot, ".dart_tool", ".cache", libTarName);
     final saveFile = File(cacheTarPath);
 
     if (force || !saveFile.existsSync()) {
       if (!saveFile.parent.existsSync()) saveFile.parent.createSync(recursive: true);
 
       String url = "https://github.com/rainyl/opencv_dart/releases/download/v$version/$libTarName";
-      print("Downloading $url");
-      final request = await HttpClient().getUrl(Uri.parse(url));
-      final response = await request.close();
-      if (response.statusCode == 200) {
-        await response.pipe(saveFile.openWrite());
-        print("Cached to $cacheTarPath");
-      } else {
-        print("Download Failed with status: ${response.statusCode}");
+      print(asInfo("Downloading $url"));
+      try {
+        final request = await HttpClient().getUrl(Uri.parse(url));
+        final response = await request.close();
+        if (response.statusCode == 200) {
+          await response.pipe(saveFile.openWrite());
+          print(asWarning("Cached to $cacheTarPath"));
+        } else {
+          print(asError("Download Failed with status: ${response.statusCode}"));
+          exit(1);
+        }
+      } catch (e) {
+        print(asError(e.toString()));
         exit(1);
       }
     } else {
-      print("Using cached $cacheTarPath");
+      print(asWarning("Using cached $cacheTarPath"));
       // Check if libs already existed, avoid double-extract
       if (Directory(extractPath)
           .listSync()
@@ -114,12 +92,12 @@ abstract class BaseSetupCommand extends Command {
               e.path.endsWith(".dylib") ||
               e.path.endsWith(".framework"))
           .any((e) => e)) {
-        print("Libs already exists in $extractPath, Skipping...");
+        print(asWarning("Libs already exists in $extractPath, Skipping..."));
         return;
       }
     }
 
-    print("Extracting to $extractPath");
+    print(asInfo("Extracting to $extractPath"));
     final tarBytes = GZipDecoder().decodeBytes(saveFile.readAsBytesSync());
     final archive = TarDecoder().decodeBytes(tarBytes);
     await extractArchiveToDisk(archive, extractPath, bufferSize: 1024 * 1024 * 10); // 10MB
@@ -127,9 +105,9 @@ abstract class BaseSetupCommand extends Command {
 
   @override
   void run() async {
-    print("opencv_dart: working for $os $arch");
-    await downloadAndExtract();
-    print("Finished");
+    print(asInfo("opencv_dart: working for $os $arch"));
+    await downloadAndExtract(arch);
+    print(asInfo("Finished"));
     exit(0);
   }
 }
@@ -145,7 +123,7 @@ class MacOsSetupCommand extends BaseSetupCommand {
     argParser.addOption(
       "arch",
       abbr: "a",
-      allowed: ["x64", "arm64"],
+      allowed: ["x86_64", "x64", "arm64"],
       mandatory: true,
     );
     argParser.addFlag("force", abbr: "f", defaultsTo: false, help: "Force download and extract");
@@ -160,7 +138,7 @@ class WindowsSetupCommand extends BaseSetupCommand {
   String get name => "windows";
 
   WindowsSetupCommand() {
-    argParser.addOption("arch", abbr: "a", allowed: ["x64"], mandatory: true);
+    argParser.addOption("arch", abbr: "a", allowed: ["x86_64", "x64"], mandatory: true);
     argParser.addFlag("force", abbr: "f", defaultsTo: false, help: "Force download and extract");
   }
 }
@@ -173,7 +151,7 @@ class LinuxSetupCommand extends BaseSetupCommand {
   String get name => "linux";
 
   LinuxSetupCommand() {
-    argParser.addOption("arch", abbr: "a", allowed: ["x64"], mandatory: true);
+    argParser.addOption("arch", abbr: "a", allowed: ["x86_64", "x64"], mandatory: true);
     argParser.addFlag("force", abbr: "f", defaultsTo: false, help: "Force download and extract");
   }
 }
@@ -207,12 +185,28 @@ class IosSetupCommand extends BaseSetupCommand {
     argParser.addOption(
       "arch",
       abbr: "a",
-      allowed: ["x64", "arm64"],
+      allowed: ["x86_64", "x64", "arm64", "os64"],
       mandatory: true,
     );
     argParser.addFlag("force", abbr: "f", defaultsTo: false, help: "Force download and extract");
   }
 }
+
+const ARCH_MAP = {
+  OS.windows: {
+    "x86_64": "x64",
+  },
+  OS.linux: {
+    "x86_64": "x64",
+  },
+  OS.macos: {
+    "x86_64": "x64",
+  },
+  OS.ios: {
+    "x86_64": "x64",
+  },
+  OS.android: {},
+};
 
 class OS {
   static const windows = "windows";
@@ -222,3 +216,9 @@ class OS {
   static const ios = "ios";
   static const macos = "macos";
 }
+
+String asInfo(String text) => '✅\x1B[32m$text\x1B[0m';
+
+String asWarning(String text) => '💡\x1B[33m$text\x1B[0m';
+
+String asError(String text) => '⛔\x1B[31m$text\x1B[0m';